ML-Week4_hw/rest_example/requirements.txt
Django==1.11.6
django-filter==1.0.4
djangorestframework==3.6.4
Markdown==2.6.9
numpy==1.13.3
pandas==0.20.3
python-dateutil==2.6.1
pytz==2017.2
scikit-learn==0.19.0
scipy==0.19.1
six==1.11.0
sklearn==0.0
